feat(entities): config store secret
